requirements.txt
# Core libraries
streamlit>=1.28.0
google-api-python-client>=2.100.0
pandas>=1.5.0
numpy>=1.24.0

# Text processing
nltk>=3.8
spacy>=3.6.0
konlpy>=0.6.0
wordcloud>=1.9.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Additional text processing
scikit-learn>=1.3.0
textblob>=0.17.0 